src/postgres_redis/requirements.txt
psycopg2-binary
types-psycopg2
pydantic
python-dotenv
redis